requirements.txt
DrissionPage==4.0.5.6
fastapi
pydantic
uvicorn
PyVirtualDisplay
pandas
bs4
google-cloud-bigquery 
python-dotenv
pyarrow
plotly
streamlit
db-dtypes
